feat: moved CSR code to separate directory
